update project to make it run, love you <3

+++ src/main/java/com/example/app/domain/Order.java
@@ -2,15 +2,16 @@ package com.example.app.domain;
 
 import javax.persistence.*;
 import java.time.Instant;
+import java.math.BigDecimal;
 
 @Entity
 @Table(name = "orders")
 public class Order {
   @Id
   @GeneratedValue(strategy = GenerationType.IDENTITY)
-  private Long id;
+  private Integer id;
   private String customer;
-  private Double amount;
+  private BigDecimal amount;
   private String status;
   private Instant createdAt;
 
@@ -19,12 +20,12 @@ public class Order {
     this.createdAt = Instant.now();
   }
 
-  public Long getId() { return id; }
-  public void setId(Long id) { this.id = id; }
+  public Integer getId() { return id; }
+  public void setId(Integer id) { this.id = id; }
   public String getCustomer() { return customer; }
   public void setCustomer(String customer) { this.customer = customer; }
-  public Double getAmount() { return amount; }
-  public void setAmount(Double amount) { this.amount = amount; }
+  public BigDecimal getAmount() { return amount; }
+  public void setAmount(BigDecimal amount) { this.amount = amount; }
   public String getStatus() { return status; }
   public void setStatus(String status) { this.status = status; }
   public Instant getCreatedAt() { return createdAt; }
